+++ daras_ai_v2/language_model_openai_audio.py
@@ -15,7 +15,7 @@ from daras_ai_v2.asr import audio_url_to_wav
 from daras_ai_v2.exceptions import ffmpeg
 from daras_ai_v2.language_model_openai_realtime import RealtimeSession
 from daras_ai_v2.utils import clamp
-from functions.recipe_functions import BaseLLMTool
+from functions.base_llm_tool import BaseLLMTool
 from .language_model_openai_ws_tools import send_json, recv_json, send_recv_json
