+++ recipes/VideoBots.py
@@ -108,8 +108,8 @@ from daras_ai_v2.vector_search import (
 )
 from functions.inbuilt_tools import get_inbuilt_tools_from_state
 from functions.models import FunctionTrigger
-from functions.recipe_functions import BaseLLMTool
-from functions.recipe_functions import (
+from functions.base_llm_tool import BaseLLMTool
+from functions.base_llm_tool import (
     get_tool_from_call,
     get_workflow_tools_from_state,
 )
